+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 2.2.0 (soon)
+# 2.2.0 ((2019-12-12)
 
 - add dry_run mode on operator
 - refactor uri module (simplier code)